requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
greenlet==3.0.3
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==3.0.1
packaging==24.1
playwright==1.47.0
pluggy==1.5.0
pyee==12.0.0
pytest==8.3.3
pytest-html==4.1.1
pytest-metadata==3.1.1
typing_extensions==4.12.2
